# swagger/tindeer.yaml
openapi: 3.0.0
info:
  title: Tindeer API
  description: Api de Rencontre Maghrébin
  version: 1.0.0
servers:
  - url: http://localhost:9080/api/v1
    description: Environnement de test Docker
  - url: http://localhost:9090/api/v1
    description: Environnement de test local
components:
  securitySchemes:
    oAuthSample:
      type: oauth2
      description: This API uses OAuth 2 with the implicit grant flow. [More info](https://api.example.com/docs/auth)
      flows:
        authorizationCode:
          authorizationUrl: http://keycloak:8080/realms/tindeer/protocol/openid-connect/auth
          tokenUrl: http://keycloak:8080/realms/tindeer/protocol/openid-connect/token
          scopes:
            profile: access profile
            email: access email
  schemas:
    MatchedUser:
      type: object
      properties:
        idMatchedUser:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        location: # Assuming you have a schema for GeoJsonPoint, adjust accordingly
          type: object
          properties:
            type:
              type: string
              enum: [Point]
            coordinates:
              type: array
              items:
                type: number
              minItems: 2
              maxItems: 2
              example: [-73.856077, 40.848447]
security:
  - oAuthSample:
      - profile
      - email
paths:
  /match/create:
    post:
      summary: Crée un nouveau profil d'utilisateur avec géolocalisation
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
              description: Address in plain text to be converted to geo-coordinates.
      responses:
        '201':
          description: Profil d'utilisateur créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchedUser'
        '400':
          description: Données invalides fournies
  /match/{id}:
    get:
      summary: Retourne le profil d'utilisateur spécifié par ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Profil d'utilisateur trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchedUser'
        '404':
          description: Profil d'utilisateur non trouvé
    put:
      summary: Met à jour le profil d'utilisateur spécifié par ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchedUser'
      responses:
        '200':
          description: Profil d'utilisateur mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchedUser'
        '400':
          description: Données invalides fournies
        '404':
          description: Profil d'utilisateur non trouvé
  /match/nearby:
    get:
      summary: Retourne les profils d'utilisateurs à proximité
      security:
        - oAuthSample:
            - profile
      responses:
        '200':
          description: Liste des profils d'utilisateurs à proximité retournée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchedUser'